Move InequalityModal.tsx to elements rather than content.
