+++ src/app/components/elements/InequalityModal.tsx
